refactor(conformance-kat): split 500-LOC helper files (closes #67)
Three test-side helper files had grown past the project's 500-LOC
guideline during the B.6 v1/v2 design arc. Each is now a directory
module with one concept per file; no semantic change.

Rust: core/tests/conformance_kat_helpers/dispatch.rs (572) →
  dispatch/{mod,open,read,inputs,lifecycle}.rs (28/84/132/144/229).
  mod.rs re-exports the entry-test surface so callers' use paths
  (run_*, assert_*) stay unchanged.

Swift: ffi/secretary-ffi-uniffi/tests/swift/conformance.swift (777) →
  4 new sibling files (ConformanceErrors/Helpers/Inputs/Assertions.swift,
  42/131/65/161 lines) + a 420-line conformance.swift carrying only
  @main ConformanceRunner. run_conformance.sh updated to pass the
  sibling list to swiftc.

Kotlin: ffi/secretary-ffi-uniffi/tests/kotlin/Conformance.kt (892) →
  same shape (4 ConformanceXyz.kt at 47/161/71/204 + Conformance.kt
  at 453 carrying fun main + dispatch when). run_conformance.sh
  updated to pass the sibling list to kotlinc.

Folds in two PR #70 review nits recorded on issue #67:
- record_uuid_hex local renamed record_uuid_str in both
  blockInputFromInputs runners (value may be wrong-length string
  when the vector uses record_uuid_bytes_hex).
- "force-cast = vector authoring contract" convention is documented
  at the top of both ConformanceInputs.swift / ConformanceInputs.kt.

Verified: 642 cargo passed / 0 failed / 10 ignored (unchanged),
clippy clean -D warnings, fmt OK, Python conformance PASS,
freshness 96/0/2 PASS, Swift conformance 22/22 PASS, Kotlin
conformance 22/22 PASS.
