+++ core/tests/conformance_kat_helpers/dispatch/read.rs
@@ -1,0 +1,132 @@
+//! read_block dispatch + record-shape assertions.
+//!
+//! `run_read_block` synthesizes `InvalidArgument` for non-16-byte
+//! `block_uuid` inputs because the bridge's `read_block` signature is
+//! `&[u8; 16]` (length is enforced by the type, not surfaced as an
+//! error variant). On the uniffi side that same length check lives in
+//! `convert_*_uuid`, so this synthesis keeps Rust + Swift + Kotlin
+//! observable-error parity.
+
+use super::super::types::{BridgeOrSyntheticErr, ExpectedRecord, OkPayload};
+
+pub fn run_read_block(
+    inputs: &serde_json::Value,
+    cached: &secretary_ffi_bridge::vault::OpenVaultOutput,
+) -> Result<secretary_ffi_bridge::record::BlockReadOutput, BridgeOrSyntheticErr> {
+    let bytes_hex = inputs
+        .get("block_uuid_hex")
+        .or_else(|| inputs.get("block_uuid_bytes_hex"))
+        .and_then(|v| v.as_str())
+        .expect("read_block inputs need block_uuid_hex or block_uuid_bytes_hex");
+    let bytes = hex::decode(bytes_hex).expect("block_uuid hex must decode");
+
+    if bytes.len() != 16 {
+        return Err(BridgeOrSyntheticErr::Synthetic {
+            variant: "InvalidArgument",
+            detail: format!("block_uuid must be exactly 16 bytes, got {}", bytes.len()),
+        });
+    }
+    let mut uuid = [0u8; 16];
+    uuid.copy_from_slice(&bytes);
+    secretary_ffi_bridge::record::read_block(&cached.identity, &cached.manifest, &uuid)
+        .map_err(BridgeOrSyntheticErr::Bridge)
+}
+
+pub fn assert_read_block_ok(
+    label: &str,
+    output: &secretary_ffi_bridge::record::BlockReadOutput,
+    expected: &OkPayload,
+) {
+    let Some(records) = &expected.records else {
+        // Vector pinned only the success shape; nothing more to check.
+        return;
+    };
+    assert_read_block_records(label, output, records);
+}
+
+pub fn assert_read_block_records(
+    label: &str,
+    output: &secretary_ffi_bridge::record::BlockReadOutput,
+    records: &[ExpectedRecord],
+) {
+    assert_eq!(
+        output.record_count(),
+        records.len(),
+        "{label}: record_count mismatch"
+    );
+    for (i, exp_rec) in records.iter().enumerate() {
+        let rec = output
+            .record_at(i)
+            .unwrap_or_else(|| panic!("{label}: record_at({i}) returned None"));
+        assert_eq!(
+            hex::encode(rec.record_uuid()),
+            exp_rec.record_uuid_hex,
+            "{label}: records[{i}].record_uuid mismatch"
+        );
+        assert_eq!(
+            rec.record_type(),
+            exp_rec.record_type,
+            "{label}: records[{i}].record_type mismatch"
+        );
+        assert_eq!(
+            rec.tags(),
+            exp_rec.tags,
+            "{label}: records[{i}].tags mismatch"
+        );
+        assert_eq!(
+            rec.field_count(),
+            exp_rec.fields.len(),
+            "{label}: records[{i}].field_count mismatch"
+        );
+        for (j, exp_field) in exp_rec.fields.iter().enumerate() {
+            let field = rec
+                .field_at(j)
+                .unwrap_or_else(|| panic!("{label}: records[{i}].field_at({j}) None"));
+            assert_eq!(
+                field.name(),
+                exp_field.name,
+                "{label}: records[{i}].fields[{j}].name mismatch"
+            );
+            match exp_field.field_type.as_str() {
+                "text" => {
+                    assert!(
+                        field.is_text(),
+                        "{label}: records[{i}].fields[{j}] expected text"
+                    );
+                    let actual = field
+                        .expose_text()
+                        .unwrap_or_else(|| panic!("{label}: expose_text returned None"));
+                    assert_eq!(
+                        &actual,
+                        exp_field
+                            .value_utf8
+                            .as_ref()
+                            .expect("text field must pin value_utf8"),
+                        "{label}: records[{i}].fields[{j}].value_utf8 mismatch"
+                    );
+                }
+                "bytes" => {
+                    assert!(
+                        field.is_bytes(),
+                        "{label}: records[{i}].fields[{j}] expected bytes"
+                    );
+                    let actual = field
+                        .expose_bytes()
+                        .unwrap_or_else(|| panic!("{label}: expose_bytes returned None"));
+                    let expected_bytes = hex::decode(
+                        exp_field
+                            .value_hex
+                            .as_ref()
+                            .expect("bytes field must pin value_hex"),
+                    )
+                    .expect("value_hex must decode");
+                    assert_eq!(
+                        actual, expected_bytes,
+                        "{label}: records[{i}].fields[{j}].value_hex mismatch"
+                    );
+                }
+                other => panic!("{label}: unknown field type '{other}'"),
+            }
+        }
+    }
+}
