+++ ffi/secretary-ffi-uniffi/tests/kotlin/Conformance.kt
@@ -7,17 +7,18 @@
 // observable output matches the pinned expectation. One PASS/FAIL line
 // per vector + a final summary.
 //
+// Helpers live in sibling files (split per issue #67 to stay under
+// the 500-LOC guideline):
+//   - ConformanceErrors.kt     — VaultException variant name + detail
+//   - ConformanceHelpers.kt    — input resolvers, hex codec, fs, after-walkers
+//   - ConformanceInputs.kt     — uniffi BlockInput / RecordInput builders
+//   - ConformanceAssertions.kt — handleOpenOk / handleVaultError / assertPostState
+//
 // Invocation: ffi/secretary-ffi-uniffi/tests/kotlin/run_conformance.sh
 
 import org.json.JSONArray
 import org.json.JSONObject
-import uniffi.secretary.BlockInput
-import uniffi.secretary.FieldInput
-import uniffi.secretary.FieldInputValue
-import uniffi.secretary.OpenVaultManifest
 import uniffi.secretary.OpenVaultOutput
-import uniffi.secretary.RecordInput
-import uniffi.secretary.UnlockedIdentity
 import uniffi.secretary.VaultException
 import uniffi.secretary.openVaultWithPassword
 import uniffi.secretary.openVaultWithRecovery
@@ -27,446 +28,6 @@ import uniffi.secretary.saveBlock
 import uniffi.secretary.shareBlock
 import uniffi.secretary.trashBlock
 import kotlin.system.exitProcess
-
-// --- Error variant name helper (mirrors the Swift vaultErrorName helper) ---
-//
-// Enumerates every variant of VaultException. If uniffi adds a new variant,
-// a new `is VaultException.XYZ` branch is needed here — the exhaustive
-// `when` will fail to compile, acting as a tripwire the same way the
-// Swift non-exhaustive-switch error does.
-
-private fun vaultExceptionVariantName(e: VaultException): String = when (e) {
-    is VaultException.WrongPasswordOrCorrupt -> "WrongPasswordOrCorrupt"
-    is VaultException.WrongMnemonicOrCorrupt -> "WrongMnemonicOrCorrupt"
-    is VaultException.InvalidMnemonic -> "InvalidMnemonic"
-    is VaultException.VaultMismatch -> "VaultMismatch"
-    is VaultException.CorruptVault -> "CorruptVault"
-    is VaultException.FolderInvalid -> "FolderInvalid"
-    is VaultException.BlockNotFound -> "BlockNotFound"
-    is VaultException.InvalidArgument -> "InvalidArgument"
-    is VaultException.SaveCryptoFailure -> "SaveCryptoFailure"
-    is VaultException.NotAuthor -> "NotAuthor"
-    is VaultException.RecipientAlreadyPresent -> "RecipientAlreadyPresent"
-    is VaultException.MissingRecipientCard -> "MissingRecipientCard"
-    is VaultException.CardDecodeFailure -> "CardDecodeFailure"
-    is VaultException.BlockUuidAlreadyLive -> "BlockUuidAlreadyLive"
-    is VaultException.BlockNotInTrash -> "BlockNotInTrash"
-}
-
-// Extract the detail string from VaultException variants that carry one.
-// Returns null for variants that carry no detail (e.g. WrongPasswordOrCorrupt).
-private fun vaultExceptionDetail(e: VaultException): String? = when (e) {
-    is VaultException.InvalidMnemonic -> e.detail
-    is VaultException.CorruptVault -> e.detail
-    is VaultException.FolderInvalid -> e.detail
-    is VaultException.InvalidArgument -> e.detail
-    is VaultException.SaveCryptoFailure -> e.detail
-    is VaultException.CardDecodeFailure -> e.detail
-    is VaultException.BlockUuidAlreadyLive -> e.detail
-    is VaultException.BlockNotInTrash -> e.detail
-    // The remaining variants carry no detail string.
-    is VaultException.WrongPasswordOrCorrupt,
-    is VaultException.WrongMnemonicOrCorrupt,
-    is VaultException.VaultMismatch,
-    is VaultException.RecipientAlreadyPresent,
-    is VaultException.NotAuthor,
-    is VaultException.MissingRecipientCard,
-    is VaultException.BlockNotFound -> null
-}
-
-// --- Input resolution helpers ---
-//
-// Mirrors the Swift resolveSource / resolveVaultDir / resolvePassword /
-// resolveMnemonic helpers in conformance.swift.
-
-private fun resolveSource(source: String, goldenVaultDir: String): ByteArray {
-    val colon = source.indexOf(':')
-    if (colon < 0) {
-        System.err.println("malformed source ref: $source")
-        exitProcess(1)
-    }
-    val file = java.io.File(goldenVaultDir, source.substring(0, colon))
-    val field = source.substring(colon + 1)
-    val text = try {
-        file.readText()
-    } catch (e: Throwable) {
-        System.err.println("failed to resolve $source: $e")
-        exitProcess(1)
-    }
-    val obj = JSONObject(text)
-    val value = obj.optString(field, null)
-    if (value == null) {
-        System.err.println("field '$field' not found in $file")
-        exitProcess(1)
-    }
-    return value.toByteArray(Charsets.UTF_8)
-}
-
-private fun resolveVaultDir(inputs: JSONObject, goldenVaultDir: String): ByteArray {
-    if (inputs.has("vault_dir")) {
-        val rel = inputs.getString("vault_dir")
-        val resolved = java.io.File(goldenVaultDir, rel).absolutePath
-        return resolved.toByteArray(Charsets.UTF_8)
-    }
-    if (inputs.has("vault_dir_literal")) {
-        return inputs.getString("vault_dir_literal").toByteArray(Charsets.UTF_8)
-    }
-    System.err.println("vector inputs missing vault_dir / vault_dir_literal")
-    exitProcess(1)
-}
-
-private fun resolvePassword(inputs: JSONObject, goldenVaultDir: String): ByteArray {
-    if (inputs.has("password_source")) return resolveSource(inputs.getString("password_source"), goldenVaultDir)
-    if (inputs.has("password_literal_utf8")) return inputs.getString("password_literal_utf8").toByteArray(Charsets.UTF_8)
-    System.err.println("vector inputs missing password_source / password_literal_utf8")
-    exitProcess(1)
-}
-
-private fun resolveMnemonic(inputs: JSONObject, goldenVaultDir: String): ByteArray {
-    if (inputs.has("mnemonic_source")) return resolveSource(inputs.getString("mnemonic_source"), goldenVaultDir)
-    if (inputs.has("mnemonic_literal_utf8")) return inputs.getString("mnemonic_literal_utf8").toByteArray(Charsets.UTF_8)
-    System.err.println("vector inputs missing mnemonic_source / mnemonic_literal_utf8")
-    exitProcess(1)
-}
-
-// Decode a lower-case hex string to a ByteArray. Exits on malformed input.
-private fun decodeHex(s: String): ByteArray {
-    if (s.length % 2 != 0) {
-        System.err.println("malformed hex (odd length): $s")
-        exitProcess(1)
-    }
-    return ByteArray(s.length / 2) { i ->
-        s.substring(i * 2, i * 2 + 2).toInt(16).toByte()
-    }
-}
-
-// Encode a ByteArray to a lower-case hex string.
-private fun encodeHex(bytes: ByteArray): String =
-    bytes.joinToString("") { "%02x".format(it) }
-
-// --- Result-arm helpers ---
-//
-// Symmetric with the Swift runner's handleOpenOk / handleVaultError.
-// Same parameter order, same assertion order. The cache is a mutable
-// reference (Kotlin maps are reference types — no inout marker needed).
-//
-// `handleVaultError` is shared by every op that throws a `VaultException`
-// (openVaultWithPassword, openVaultWithRecovery, readBlock) — the
-// variant + detail_contains contract is uniform across them.
-
-private fun handleOpenOk(
-    out: OpenVaultOutput,
-    expected: JSONObject,
-    name: String,
-    kind: String,
-    cache: MutableMap<String, OpenVaultOutput>,
-    check: (Boolean, String, String) -> Boolean,
-) {
-    if (kind != "ok") {
-        check(false, name, "expected err, got ok")
-        return
-    }
-    // Aggregate sub-check results so we only cache on full success.
-    // Matches the Rust replay (assert_open_ok panics on mismatch and
-    // cache.insert never runs) — chained read_block vectors then
-    // report "predecessor did not produce a cacheable Ok" instead of
-    // running against a vault whose pinned metadata didn't match.
-    var allOk = true
-    expected.optString("display_name", null)?.let { wantDisplay ->
-        if (!check(out.identity.displayName() == wantDisplay, name,
-                "display_name mismatch (got '${out.identity.displayName()}', want '$wantDisplay')")) {
-            allOk = false
-        }
-    }
-    if (expected.has("block_count")) {
-        val wantBc = expected.getInt("block_count")
-        if (!check(out.manifest.blockCount().toInt() == wantBc, name,
-                "block_count mismatch (got ${out.manifest.blockCount()}, want $wantBc)")) {
-            allOk = false
-        }
-    }
-    expected.optString("block_uuid_hex", null)?.let { wantUuid ->
-        val summaries = out.manifest.blockSummaries()
-        if (summaries.isNotEmpty()) {
-            if (!check(encodeHex(summaries[0].blockUuid) == wantUuid, name,
-                    "block_uuid mismatch (got '${encodeHex(summaries[0].blockUuid)}', want '$wantUuid')")) {
-                allOk = false
-            }
-        } else {
-            check(false, name, "manifest has no blocks but block_uuid_hex pinned")
-            allOk = false
-        }
-    }
-    if (allOk) {
-        cache[name] = out
-    }
-}
-
-private fun handleVaultError(
-    e: VaultException,
-    expected: JSONObject,
-    name: String,
-    kind: String,
-    check: (Boolean, String, String) -> Boolean,
-) {
-    if (kind != "err") {
-        check(false, name, "expected ok, got err: $e")
-        return
-    }
-    val wantVariant = expected.optString("variant", "")
-    val gotVariant = vaultExceptionVariantName(e)
-    check(gotVariant == wantVariant, name, "variant mismatch (got $gotVariant, expected $wantVariant)")
-    expected.optString("detail_contains", null)?.let { needle ->
-        val detail = vaultExceptionDetail(e) ?: ""
-        check(detail.contains(needle), name, "detail '$detail' missing '$needle'")
-    }
-}
-
-// --- v2 lifecycle helpers ---
-
-private fun recursiveCopy(src: java.nio.file.Path, dst: java.nio.file.Path) {
-    java.nio.file.Files.walk(src).use { stream ->
-        stream.forEach { entry ->
-            val rel = src.relativize(entry)
-            val target = dst.resolve(rel.toString())
-            if (java.nio.file.Files.isDirectory(entry)) {
-                java.nio.file.Files.createDirectories(target)
-            } else {
-                java.nio.file.Files.copy(entry, target, java.nio.file.StandardCopyOption.REPLACE_EXISTING)
-            }
-        }
-    }
-}
-
-private fun cleanupTempVault(tmp: java.nio.file.Path) {
-    if (!java.nio.file.Files.exists(tmp)) return
-    java.nio.file.Files.walk(tmp).use { stream ->
-        stream.sorted(java.util.Comparator.reverseOrder()).forEach {
-            try {
-                java.nio.file.Files.deleteIfExists(it)
-            } catch (_: java.io.IOException) { /* best-effort */ }
-        }
-    }
-}
-
-private fun readContactCardBytes(vaultDir: java.nio.file.Path, userUuidHex: String): ByteArray {
-    require(userUuidHex.length == 32) { "userUuidHex must be 32 chars" }
-    val h = userUuidHex
-    val hyphenated = "${h.substring(0, 8)}-${h.substring(8, 12)}-${h.substring(12, 16)}-${h.substring(16, 20)}-${h.substring(20, 32)}.card"
-    val path = vaultDir.resolve("contacts").resolve(hyphenated)
-    return java.nio.file.Files.readAllBytes(path)
-}
-
-private fun findWritableDir(
-    start: String,
-    writableVaultDirs: Map<String, java.nio.file.Path>,
-    vectors: JSONArray,
-): java.nio.file.Path? {
-    var current = start
-    // Bounded by vectors.length() — an authoring-error `after:` cycle would
-    // otherwise hang. Fail loudly so the cycle is fixable, not silent.
-    for (step in 0..vectors.length()) {
-        writableVaultDirs[current]?.let { return it }
-        var parentAfter: String? = null
-        for (i in 0 until vectors.length()) {
-            val v = vectors.getJSONObject(i)
-            if (v.optString("name", null) == current) {
-                parentAfter = if (v.has("after")) v.getString("after") else null
-                break
-            }
-        }
-        current = parentAfter ?: return null
-    }
-    error("after-chain cycle detected starting at '$start' (depth exceeded vectors.length())")
-}
-
-private fun findCacheAncestorName(
-    start: String,
-    cache: Map<String, OpenVaultOutput>,
-    vectors: JSONArray,
-): String? {
-    var current = start
-    // Cycle guard: see findWritableDir.
-    for (step in 0..vectors.length()) {
-        if (cache.containsKey(current)) return current
-        var parentAfter: String? = null
-        for (i in 0 until vectors.length()) {
-            val v = vectors.getJSONObject(i)
-            if (v.optString("name", null) == current) {
-                parentAfter = if (v.has("after")) v.getString("after") else null
-                break
-            }
-        }
-        current = parentAfter ?: return null
-    }
-    error("after-chain cycle detected starting at '$start' (depth exceeded vectors.length())")
-}
-
-/**
- * Build a BlockInput from the JSON `inputs` dict. Mirrors
- * `block_input_from_inputs` in core/tests/conformance_kat_helpers/dispatch.rs
- * and `blockInputFromInputs` in the Swift runner.
- *
- * `record_uuid_hex` is the happy-path key (always 16 bytes); a vector may
- * instead provide `record_uuid_bytes_hex` with a wrong-length value to
- * exercise uniffi's namespace-layer length check inside
- * `convert_record_input`, which surfaces `VaultException.InvalidArgument`
- * symmetrically with the wrong-length device_uuid / block_uuid paths.
- */
-private fun blockInputFromInputs(inputs: JSONObject): BlockInput {
-    val blockUuidHex = inputs.getString("block_uuid_hex")
-    val blockName = if (inputs.has("block_name")) inputs.getString("block_name") else ""
-    val recordsArr = if (inputs.has("records")) inputs.getJSONArray("records") else JSONArray()
-    val records = (0 until recordsArr.length()).map { i ->
-        val rec = recordsArr.getJSONObject(i)
-        val recordUuidHex = if (rec.has("record_uuid_hex")) rec.getString("record_uuid_hex") else rec.getString("record_uuid_bytes_hex")
-        val fieldsArr = if (rec.has("fields")) rec.getJSONArray("fields") else JSONArray()
-        val fields = (0 until fieldsArr.length()).map { j ->
-            val f = fieldsArr.getJSONObject(j)
-            val fname = f.getString("name")
-            val ftype = f.getString("type")
-            val value: FieldInputValue = when (ftype) {
-                "text" -> FieldInputValue.Text(f.getString("value_utf8"))
-                "bytes" -> FieldInputValue.Bytes(decodeHex(f.getString("value_hex")))
-                else -> error("unknown field type: $ftype")
-            }
-            FieldInput(fname, value)
-        }
-        RecordInput(decodeHex(recordUuidHex), fields)
-    }
-    return BlockInput(decodeHex(blockUuidHex), blockName, records)
-}
-
-// Returns 16-byte UUID or null if input is the wrong length / missing.
-private fun uuidFromInputs(inputs: JSONObject, primary: String, bytes: String): ByteArray? {
-    val raw = when {
-        inputs.has(primary) -> inputs.getString(primary)
-        inputs.has(bytes) -> inputs.getString(bytes)
-        else -> return null
-    }
-    val data = decodeHex(raw)
-    return if (data.size == 16) data else null
-}
-
-private fun assertPostState(
-    name: String,
-    identity: UnlockedIdentity,
-    manifest: OpenVaultManifest,
-    expected: JSONObject,
-    check: (Boolean, String, String) -> Boolean,
-) {
-    val postState = expected.optJSONObject("post_state") ?: return
-    if (postState.has("block_count")) {
-        val bc = postState.getInt("block_count")
-        check(manifest.blockCount().toInt() == bc, name, "post_state.block_count mismatch (got ${manifest.blockCount()}, want $bc)")
-    }
-    var roundTripUuid: ByteArray? = null
-    if (postState.has("find_block_uuid_hex") && !postState.isNull("find_block_uuid_hex")) {
-        val hexStr = postState.getString("find_block_uuid_hex")
-        val uuidData = decodeHex(hexStr)
-        val summary = manifest.findBlock(uuidData)
-        if (summary == null) {
-            check(false, name, "post_state.find_block_uuid_hex=$hexStr not in manifest")
-        } else {
-            check(summary.blockUuid.contentEquals(uuidData), name, "post_state.find_block returned wrong uuid")
-            roundTripUuid = uuidData
-        }
-    }
-    if (postState.has("recipient_count")) {
-        val rc = postState.getInt("recipient_count")
-        val uuid = roundTripUuid
-        if (uuid == null) {
-            check(false, name, "post_state.recipient_count requires find_block_uuid_hex")
-            return
-        }
-        val summary = manifest.findBlock(uuid)
-        if (summary == null) {
-            check(false, name, "recipient_count: block not findable")
-            return
-        }
-        check(summary.recipientUuids.size == rc, name, "post_state.recipient_count mismatch (got ${summary.recipientUuids.size}, want $rc)")
-    }
-    if (postState.has("read_block")) {
-        val readPin = postState.getJSONObject("read_block")
-        val pinnedRecords = readPin.getJSONArray("records")
-        val uuid = roundTripUuid
-        if (uuid == null) {
-            check(false, name, "post_state.read_block requires find_block_uuid_hex")
-            return
-        }
-        try {
-            val output = readBlock(identity, manifest, uuid)
-            try {
-                check(output.recordCount().toInt() == pinnedRecords.length(), name, "post_state.read_block.record_count mismatch")
-                for (ri in 0 until pinnedRecords.length()) {
-                    val expRec = pinnedRecords.getJSONObject(ri)
-                    val rec = output.recordAt(ri.toULong())
-                    if (rec == null) {
-                        check(false, name, "post_state.read_block.record_at($ri) returned null")
-                        continue
-                    }
-                    try {
-                        expRec.optString("record_uuid_hex", null)?.let { wantUuid ->
-                            check(encodeHex(rec.recordUuid()) == wantUuid, name, "records[$ri].record_uuid mismatch")
-                        }
-                        expRec.optString("record_type", null)?.let { wantType ->
-                            check(rec.recordType() == wantType, name, "records[$ri].record_type mismatch")
-                        }
-                        if (expRec.has("tags")) {
-                            val expTags = expRec.getJSONArray("tags")
-                            val expTagsList = (0 until expTags.length()).map { expTags.getString(it) }
-                            check(rec.tags() == expTagsList, name, "records[$ri].tags mismatch")
-                        }
-                        if (expRec.has("fields")) {
-                            val expFields = expRec.getJSONArray("fields")
-                            check(rec.fieldCount().toInt() == expFields.length(), name, "records[$ri].field_count mismatch")
-                            for (fi in 0 until expFields.length()) {
-                                val expF = expFields.getJSONObject(fi)
-                                val fh = rec.fieldAt(fi.toULong())
-                                if (fh == null) {
-                                    check(false, name, "records[$ri].field_at($fi) returned null")
-                                    continue
-                                }
-                                try {
-                                    expF.optString("name", null)?.let { wantName ->
-                                        check(fh.name() == wantName, name, "records[$ri].fields[$fi].name mismatch")
-                                    }
-                                    expF.optString("type", null)?.let { ftype ->
-                                        when (ftype) {
-                                            "text" -> {
-                                                check(fh.isText(), name, "records[$ri].fields[$fi] expected text")
-                                                expF.optString("value_utf8", null)?.let { wantVal ->
-                                                    check(fh.exposeText() == wantVal, name, "records[$ri].fields[$fi].value_utf8 mismatch")
-                                                }
-                                            }
-                                            "bytes" -> {
-                                                check(fh.isBytes(), name, "records[$ri].fields[$fi] expected bytes")
-                                                expF.optString("value_hex", null)?.let { wantHex ->
-                                                    val actual = encodeHex(fh.exposeBytes() ?: ByteArray(0))
-                                                    check(actual == wantHex, name, "records[$ri].fields[$fi].value_hex mismatch")
-                                                }
-                                            }
-                                        }
-                                    }
-                                } finally {
-                                    fh.destroy()
-                                }
-                            }
-                        }
-                    } finally {
-                        rec.destroy()
-                    }
-                }
-            } finally {
-                output.destroy()
-            }
-        } catch (e: Throwable) {
-            check(false, name, "post_state.read_block threw $e")
-        }
-    }
-}
-
-// --- Main ---
 
 fun main() {
     // --- Environment ---
