+++ ffi/secretary-ffi-uniffi/tests/kotlin/ConformanceInputs.kt
@@ -1,0 +1,71 @@
+// JSON → uniffi input builders for v2 write ops.
+//
+// Mirrors core/tests/conformance_kat_helpers/dispatch/inputs.rs (Rust)
+// and ConformanceInputs.swift.
+//
+// Convention: getString-without-default is a vector-authoring contract.
+// If a vector reaches one of these helpers without the expected key
+// (e.g. records[i].fields[j] missing `name`), org.json will throw
+// JSONException — the trap is intentional. It means the vector itself
+// is malformed, not that the runtime input happened to be the wrong
+// type. Per-field defensive `optString(key, null) ?: error("...")`
+// would obscure the authoring error behind a generic "vector failed"
+// message. The KAT JSON is generated by Rust + hand-edited; the test
+// runner is the validator of last resort, so authoring errors should
+// crash loudly with the variable name + line number visible in the
+// stack trace.
+
+import org.json.JSONArray
+import org.json.JSONObject
+import uniffi.secretary.BlockInput
+import uniffi.secretary.FieldInput
+import uniffi.secretary.FieldInputValue
+import uniffi.secretary.RecordInput
+
+// Returns 16-byte UUID or null if input is the wrong length / missing.
+internal fun uuidFromInputs(inputs: JSONObject, primary: String, bytes: String): ByteArray? {
+    val raw = when {
+        inputs.has(primary) -> inputs.getString(primary)
+        inputs.has(bytes) -> inputs.getString(bytes)
+        else -> return null
+    }
+    val data = decodeHex(raw)
+    return if (data.size == 16) data else null
+}
+
+/**
+ * Build a BlockInput from the JSON `inputs` dict. Mirrors
+ * `block_input_from_inputs` in core/tests/conformance_kat_helpers/dispatch/inputs.rs
+ * and `blockInputFromInputs` in the Swift runner.
+ *
+ * `record_uuid_hex` is the happy-path key (always 16 bytes); a vector may
+ * instead provide `record_uuid_bytes_hex` with a wrong-length value to
+ * exercise uniffi's namespace-layer length check inside
+ * `convert_record_input`, which surfaces `VaultException.InvalidArgument`
+ * symmetrically with the wrong-length device_uuid / block_uuid paths.
+ * The locally-bound `recordUuidStr` name reflects that the value may be
+ * a wrong-length string rather than a canonical 16-byte hex.
+ */
+internal fun blockInputFromInputs(inputs: JSONObject): BlockInput {
+    val blockUuidHex = inputs.getString("block_uuid_hex")
+    val blockName = if (inputs.has("block_name")) inputs.getString("block_name") else ""
+    val recordsArr = if (inputs.has("records")) inputs.getJSONArray("records") else JSONArray()
+    val records = (0 until recordsArr.length()).map { i ->
+        val rec = recordsArr.getJSONObject(i)
+        val recordUuidStr = if (rec.has("record_uuid_hex")) rec.getString("record_uuid_hex") else rec.getString("record_uuid_bytes_hex")
+        val fieldsArr = if (rec.has("fields")) rec.getJSONArray("fields") else JSONArray()
+        val fields = (0 until fieldsArr.length()).map { j ->
+            val f = fieldsArr.getJSONObject(j)
+            val fname = f.getString("name")
+            val ftype = f.getString("type")
+            val value: FieldInputValue = when (ftype) {
+                "text" -> FieldInputValue.Text(f.getString("value_utf8"))
+                "bytes" -> FieldInputValue.Bytes(decodeHex(f.getString("value_hex")))
+                else -> error("unknown field type: $ftype")
+            }
+            FieldInput(fname, value)
+        }
+        RecordInput(decodeHex(recordUuidStr), fields)
+    }
+    return BlockInput(decodeHex(blockUuidHex), blockName, records)
+}
