+++ ffi/secretary-ffi-uniffi/tests/swift/ConformanceInputs.swift
@@ -1,0 +1,65 @@
+// JSON → uniffi input builders for v2 write ops.
+//
+// Mirrors core/tests/conformance_kat_helpers/dispatch/inputs.rs (Rust)
+// and ConformanceInputs.kt (Kotlin).
+//
+// Convention: cast-to-required (`as!`) is a vector-authoring contract.
+// If a vector reaches one of these helpers without the expected key
+// (e.g. records[i].fields[j] missing `name`), the trap is intentional —
+// it means the vector itself is malformed, not that the runtime input
+// happened to be the wrong type. Per-field defensive `guard let … else`
+// would obscure the authoring error behind a generic "vector failed"
+// message. The KAT JSON is generated by Rust + hand-edited; the test
+// runner is the validator of last resort, so authoring errors should
+// crash loudly with the variable name + line number visible in the
+// stack trace.
+
+import Foundation
+
+/// uuid extraction matching Rust's uuid_from_inputs. Returns nil for
+/// non-16-byte input (caller surfaces synthesized InvalidArgument).
+func uuidFromInputs(_ inputs: [String: Any], primary: String, bytes: String) -> Data? {
+    let raw: String
+    if let s = inputs[primary] as? String { raw = s }
+    else if let s = inputs[bytes] as? String { raw = s }
+    else { return nil }
+    let data = decodeHex(raw)
+    return data.count == 16 ? data : nil
+}
+
+/// Build a BlockInput from the JSON `inputs` dict. Mirrors
+/// `block_input_from_inputs` in core/tests/conformance_kat_helpers/dispatch/inputs.rs.
+///
+/// `record_uuid_hex` is the happy-path key (always 16 bytes); a vector may
+/// instead provide `record_uuid_bytes_hex` with a wrong-length value to
+/// exercise uniffi's namespace-layer `uuid_from_vec` length check (the
+/// bridge's `RecordInput.record_uuid` is type-bounded to `[u8; 16]` in
+/// Rust, but on the uniffi side it's a `Data`/`ByteArray` whose length
+/// is validated inside `convert_record_input`, surfacing
+/// `VaultError.InvalidArgument` symmetrically with wrong-length
+/// device_uuid / block_uuid). The locally-bound `recordUuidStr` name
+/// reflects that the value may be a wrong-length string rather than a
+/// canonical 16-byte hex.
+func blockInputFromInputs(_ inputs: [String: Any]) -> BlockInput {
+    let blockUuidHex = inputs["block_uuid_hex"] as! String
+    let blockName = (inputs["block_name"] as? String) ?? ""
+    let records: [RecordInput] = (inputs["records"] as? [[String: Any]] ?? []).map { rec in
+        let recordUuidStr = (rec["record_uuid_hex"] as? String) ?? (rec["record_uuid_bytes_hex"] as! String)
+        let fields: [FieldInput] = (rec["fields"] as? [[String: Any]] ?? []).map { f in
+            let fname = f["name"] as! String
+            let ftype = f["type"] as! String
+            let value: FieldInputValue
+            switch ftype {
+            case "text":
+                value = .text(text: f["value_utf8"] as! String)
+            case "bytes":
+                value = .bytes(data: decodeHex(f["value_hex"] as! String))
+            default:
+                fatalError("unknown field type: \(ftype)")
+            }
+            return FieldInput(name: fname, value: value)
+        }
+        return RecordInput(recordUuid: decodeHex(recordUuidStr), fields: fields)
+    }
+    return BlockInput(blockUuid: decodeHex(blockUuidHex), blockName: blockName, records: records)
+}
